requirements.txt
prisma
APScheduler
numpy
python-dotenv
geopy